SQL_search/requirements.txt
python-dotenv==1.0.0
mysql-connector-python==8.2.0
openai==0.28
datetime==5.4
functools==0.5
unittest2==1.1.0
logging==0.5.1.2
abc==0.0.1
hashlib==1.0.3
typing==3.7.4.3
pyodbc==4.0.39